GLUE-248 : CHORE - 서비스 도메인 분리
