+++ src/main/java/com/justdo/plug/post/domain/post/service/PostService.java
@@ -1,13 +1,12 @@
 package com.justdo.plug.post.domain.post.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.justdo.plug.post.domain.hashtag.service.HashtagService;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.justdo.plug.post.domain.post.Post;
 import com.justdo.plug.post.domain.post.dto.PostRequestDto;
 import com.justdo.plug.post.domain.post.dto.PostResponseDto;
 import com.justdo.plug.post.domain.post.repository.PostRepository;
-import com.justdo.plug.post.domain.posthashtag.PostHashtag;
-import com.justdo.plug.post.domain.posthashtag.service.PostHashtagService;
 import com.justdo.plug.post.global.exception.ApiException;
 import com.justdo.plug.post.global.response.code.status.ErrorStatus;
 import jakarta.transaction.Transactional;
@@ -15,9 +14,6 @@ import lombok.RequiredArgsConstructor;
 import org.json.JSONException;
 import org.springframework.stereotype.Service;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.ArrayList;
 import java.util.List;
 
 
@@ -26,8 +22,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
-    private final HashtagService hashtagService;
-    private final PostHashtagService postHashtagService;
+
 
 
     // BLOG001: 게시글 리스트 조회
@@ -59,39 +54,6 @@ public class PostService {
             throw new ApiException(ErrorStatus._BLOG_NOT_FOUND);
         }
         return posts;
-    }
-
-    // BLOG007: 해시태그 값 추출
-    public List<String> getHashtags(Long memberId) {
-        // 멤버 아이디에 해당하는 포스트를 가져온다.
-        List<Post> memberPosts = postRepository.findByMemberId(memberId);
-
-        // 멤버 아이디에 해당하는 포스트가 없는 경우
-        if (memberPosts.isEmpty()) {
-            throw new ApiException(ErrorStatus._NO_HASHTAGS);
-        }
-
-        // 멤버 아이디에 해당하는 포스트의 아이디만 추출하여 반환
-        List<Long> postIds = memberPosts.stream()
-                .map(Post::getId)
-                .toList();
-
-        List<String> hashtagNames = new ArrayList<>();
-
-        // 각 포스트별로 해당하는 해시태그 아이디를 추출하여 저장
-        for (Long postId : postIds) {
-            List<PostHashtag> postHashtags;
-            postHashtags = postHashtagService.getPostHashtags(postId);
-
-            for (PostHashtag postHashtag : postHashtags) {
-                // 아이디에서 해시태그 명으로 변경 후 리스트에 저장
-                String hashtagName = hashtagService.getHashtagNameById(postHashtag.getHashtagId());
-                hashtagNames.add(hashtagName);
-            }
-        }
-
-
-        return hashtagNames;
     }
 
     // BlOG008: 게시글의 글만 조회하기
@@ -142,38 +104,6 @@ public class PostService {
         return extractedTexts.toString().trim();
     }
 
-    // BLOG009: 블로그 아이디로 해시태그 추출하기
-    public List<String> getHashtagsBlog(Long blogId){
 
-        // 블로그 아이디에 해당하는 포스트를 가져온다.
-        List<Post> blogPosts = postRepository.findByBlogId(blogId);
-
-        // 블로그 아이디에 해당하는 포스트가 없는 경우
-        if (blogPosts.isEmpty()) {
-            throw new ApiException(ErrorStatus._NO_HASHTAGS);
-        }
-
-        // 블로그 아이디에 해당하는 포스트의 아이디만 추출하여 반환
-        List<Long> postIds = blogPosts.stream()
-                .map(Post::getId)
-                .toList();
-
-        List<String> hashtagNames = new ArrayList<>();
-
-        // 각 포스트별로 해당하는 해시태그 아이디를 추출하여 저장
-        for (Long postId : postIds) {
-            List<PostHashtag> postHashtags;
-            postHashtags = postHashtagService.getPostHashtags(postId);
-
-            for (PostHashtag postHashtag : postHashtags) {
-                // 아이디에서 해시태그 명으로 변경 후 리스트에 저장
-                String hashtagName = hashtagService.getHashtagNameById(postHashtag.getHashtagId());
-                hashtagNames.add(hashtagName);
-            }
-        }
-
-
-        return hashtagNames;
-    }
 
 }
